refactor(generator): split main.rs into modules
